moved to the mysql

+++ src/main/java/com/devo/product/domain/BaseEntity.java
@@ -5,7 +5,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.GenericGenerator;
 import org.hibernate.annotations.UpdateTimestamp;
 
 import java.sql.Timestamp;
@@ -17,33 +16,20 @@ import java.sql.Timestamp;
 public class BaseEntity {
 
     @Id
-    @GeneratedValue(strategy=GenerationType.AUTO)
-//    @GeneratedValue(generator = "UUID")
-//    @GenericGenerator(
-//            name = "UUID",
-//            strategy = "org.hibernate.id.UUIDGenerator"
-//    )
-    //@Column(length = 36, columnDefinition = "varchar", updatable = false, nullable = false)
-    //@Column(updatable = false, nullable = false)
-    private Integer id;
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
 
     @Version
     private Long version;
 
     @CreationTimestamp
     @Column(updatable = false)
-    //@Temporal(TemporalType.TIMESTAMP)
-    //@Column(name = "create_date", updatable = false)
     private Timestamp createdDate;
 
     @UpdateTimestamp
-    // @Column
-    //@Temporal(TemporalType.TIMESTAMP)
-    //@Column(name = "modify_date")
     private Timestamp lastModifiedDate;
 
-
-    public BaseEntity(Integer id, Long version, Timestamp createdDate, Timestamp lastModifiedDate) {
+    public BaseEntity(Long id, Long version, Timestamp createdDate, Timestamp lastModifiedDate) {
         this.id = id;
         this.version = version;
         this.createdDate = createdDate;
