+++ src/main/java/com/devo/product/domain/ProductEntity.java
@@ -34,10 +34,17 @@ public class ProductEntity extends BaseEntity {
     @Fetch(FetchMode.JOIN)
     private Set<ProductInventoryEntity> productInventoryEntity = new HashSet<>();
 
-
     @Builder
-    public ProductEntity(Integer id, Long version, Timestamp createdDate, Timestamp lastModifiedDate,
-                         String title, ProductTypeEnum type, BigDecimal price, Set<ProductInventoryEntity> productInventoryEntity) {
+    public ProductEntity(
+            Long id,
+            Long version,
+            Timestamp createdDate,
+            Timestamp lastModifiedDate,
+            String title,
+            ProductTypeEnum type,
+            BigDecimal price,
+            Set<ProductInventoryEntity> productInventoryEntity
+    ) {
         super(id, version, createdDate, lastModifiedDate);
         this.title = title;
         this.type = type;
