+++ src/main/java/com/devo/product/domain/ProductOrderLineEntity.java
@@ -21,7 +21,7 @@ public class ProductOrderLineEntity extends BaseEntity {
     private Integer quantityAllocated = 0;
 
     @Builder
-    public ProductOrderLineEntity(Integer id, Long version, Timestamp createdDate, Timestamp lastModifiedDate, ProductEntity productEntity, Integer quantityAllocated) {
+    public ProductOrderLineEntity(Long id, Long version, Timestamp createdDate, Timestamp lastModifiedDate, ProductEntity productEntity, Integer quantityAllocated) {
         super(id, version, createdDate, lastModifiedDate);
         this.productEntity = productEntity;
         this.quantityAllocated = quantityAllocated;
