+++ src/main/java/com/devo/product/services/CustomerService.java
@@ -3,5 +3,5 @@ package com.devo.product.services;
 import java.util.List;
 
 public interface CustomerService {
-  List<Integer> getAllCustomerIds();
+  List<Long> getAllCustomerIds();
 }
